requirements.txt
google-generativeai
beautifulsoup4
python-docx
Pillow
streamlit